config: validate the set wake up interval
The wake up interval shoud be treated like other config parameters.
the config command already read the wake up interval, but it was only for the log file. now it should be validated. if the check fails, the configuration will be marked as failed in the log. if the command was queued, it will be requeued.