new command cancel in ClientTCP
command cancel stops the current command if  necessary, but do not stop the program if there is no current command.
It works only in ClientTCP. in the hyper console the command would not be executed until the current command completes. But for test automation it is exactly what we need.